+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file. The format 
 
 ## [Unreleased]
 
+### Changed
+
+- Handle legacy GeoGebra applets more gracefully
+
 ### Fixed
 
 - Fix slate plugin